refactor: 시간 관련 공통 유틸 분리
