+++ frontend/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./formatDate";
 export * from "./pathToLayoutKey";
 export * from "./validation";
-export * from './formatErrMsg'
+export * from "./formatErrMsg";
+export * from "./calculateTime";
